+++ kermit/k95/ckoads.h
@@ -1,0 +1,10 @@
+/*  C K O A D S . H  --  Header for ADDS Emulation  */
+
+/*
+  Author: David Goodwin <david@zx.net.nz>
+
+  Copyright (C) 2024 David Goodwin
+*/
+_PROTOTYP( void addsascii, ( int ) ) ;
+
+