+++ kermit/k95/ckuus5.c
@@ -5315,6 +5315,7 @@ static int nshokey = (sizeof(shokeytab) / sizeof(struct keytab));
 #define SHKEYDEF TT_MAX+7
 struct keytab shokeymtab[] = {
     "aaa",       TT_AAA,     CM_INV,    /* AnnArbor */
+    "adds25",    TT_REGENT25,0,         /* ADDS Regent 25 */
     "adm3a",     TT_ADM3A,   0,         /* LSI ADM-3A */
     "adm5",      TT_ADM5,    0,         /* LSI ADM-5 */
     "aixterm",   TT_AIXTERM, 0,         /* IBM AIXterm */
@@ -5347,6 +5348,7 @@ struct keytab shokeymtab[] = {
     "meta",      TT_KBM_METAESC,   0,   /* Meta sends ESC mode (subset of emacs mode) */
     "qansi",     TT_QANSI,   0,         /* QNX ANSI */
     "qnx",       TT_QNX,     0,         /* QNX */
+    "regent25",  TT_REGENT25,CM_INV,    /* ADDS Regent 25 */
     "russian",   TT_KBM_RUSSIAN, 0,     /* Russian mode */
     "scoansi",   TT_SCOANSI, 0,         /* SCO ANSI */
     "sni-97801", TT_97801,   0,         /* Sinix 97801 */
@@ -6118,9 +6120,13 @@ shotrm() {
     int lines = 0;
     extern int colorpalette;
 #ifdef KUI
+    extern int tt_bell_flash;
+#endif /* KUI */
+#ifdef KUI
     extern CKFLOAT tt_linespacing[];
     extern int tt_cursor_blink;
 #endif /* KUI */
+   char bell[64] = "";
 #ifdef PCFONTS
     int i;
     char *font;
@@ -6220,21 +6226,26 @@ shotrm() {
            showoff(tt_answer),"response",answerback);
     switch (tt_bell) {
       case XYB_NONE:
-        s = "none";
+        ckstrncat(bell,"none",64);
         break;
       case XYB_VIS:
-        s= "visible";
+        ckstrncat(bell,"visible",64);
         break;
       case XYB_AUD | XYB_BEEP:
-        s="beep";
+        ckstrncat(bell,"beep",64);
         break;
       case XYB_AUD | XYB_SYS:
-        s="system sounds";
+        ckstrncat(bell,"system sounds",64);
         break;
       default:
-        s="(unknown)";
+        ckstrncat(bell,"(unknown)",64);
     }
-    printf(" %19s: %-13s  %13s: %-15s\n","Bell",s,
+#ifdef KUI
+    if (tt_bell_flash) {
+      ckstrncat(bell,", flash",64);
+    }
+#endif /* KUI */
+    printf(" %19s: %-21s  %5s: %-15s\n","Bell",bell,
            "Wrap",showoff(tt_wrap));
     if (++lines > cmd_rows - 3) { if (!askmore()) return; else lines = 0; }
     printf(" %19s: %-13s  %13s: %-15s\n","Autopage",showoff(wy_autopage),