+++ kermit/k95/ckuus7.c
@@ -1291,6 +1291,11 @@ extern int autodl, adl_err, adl_ask;
 struct keytab beltab[] = {              /* Terminal bell mode */
 #ifdef OS2
     { "audible", XYB_AUD,  0 },
+#ifdef KUI
+    { "flash-window", XYB_FLASH, 0},
+#else
+    { "flash-window", XYB_FLASH, CM_INV },
+#endif /* KUI */
     { "none",    XYB_NONE, 0 },
 #else
     { "audible", XYB_AUD,  CM_INV },
@@ -1623,6 +1628,7 @@ int ncursors = 4;
 
 struct keytab ttyptab[] = {
     { "aaa",      TT_AAA,     CM_INV },     /* AnnArbor */
+    { "adds25",   TT_REGENT25,0 },          /* ADDS Regent 25 */
     { "adm3a",    TT_ADM3A,   0 },          /* LSI ADM-3A */
     { "adm5",     TT_ADM5,    0 },          /* LSI ADM-5 */
     { "aixterm",  TT_AIXTERM, 0 },          /* IBM AIXterm */
@@ -1652,6 +1658,7 @@ struct keytab ttyptab[] = {
     { "linux",    TT_LINUX,   0 },          /* Linux */
     { "qansi",    TT_QANSI,   0 },          /* QNX ANSI */
     { "qnx",      TT_QNX,     0 },          /* QNX Console */
+    { "regent25", TT_REGENT25,CM_INV },     /* ADDS Regent 25 */
     { "scoansi",  TT_SCOANSI, 0 },          /* SCO ANSI */
     { "sni-97801",TT_97801,   0 },          /* SNI 97801 */
     { "sun",      TT_SUN,     0 },          /* SUN Console */
@@ -1713,6 +1720,7 @@ int nttyp = (sizeof(ttyptab) / sizeof(struct keytab));
 
 struct keytab ttkeytab[] = {
     { "aaa",       TT_AAA,        CM_INV },        /* AnnArbor */
+    { "adds25",    TT_REGENT25,   0 },             /* ADDS Regent 25 */
     { "adm3a",     TT_ADM3A,      0 },             /* LSI ADM-3A */
     { "adm5",      TT_ADM5,       0 },             /* LSI ADM-5 */
     { "aixterm",   TT_AIXTERM,    0 },             /* IBM AIXterm */
@@ -1744,6 +1752,7 @@ struct keytab ttkeytab[] = {
     { "meta",      TT_KBM_METAESC, 0, },           /* Meta sends ESC mode (subset of emacs mode) */
     { "qansi",     TT_QANSI,      0 },             /* QNX ANSI */
     { "qnx",       TT_QNX,        0 },             /* QNX */
+    { "regent25",  TT_REGENT25,   CM_INV },        /* ADDS Regent 25 */
     { "russian",   TT_KBM_RUSSIAN,0 },             /* Russian mode */
     { "scoansi",   TT_SCOANSI,    0 },             /* SCO ANSI */
     { "sni-97801", TT_97801,      0 },             /* SNI 97801 */
@@ -6737,6 +6746,17 @@ setbell() {
         tt_bell = 1;
 #endif /* OS2 */
         break;
+#ifdef OS2
+      case XYB_FLASH: {
+#ifdef KUI
+        extern int user_bell_flash, tt_bell_flash;
+        if ((x = cmkey(onoff,2,"","on",xxstring)) < 0) return(x);
+        if ((z = cmcfm()) < 0) return(z);
+        user_bell_flash = tt_bell_flash = x;
+#endif /* KUI */
+      }
+      break;
+#endif /* OS2 */
     }
     return(1);
 }