+++ kermit/k95/ckuusr.h
@@ -1063,11 +1063,19 @@ struct stringint {			/* String and (wide) integer */
 #define     TT_ADM3A   40       /*    LSI ADM 3A */
 #define     TT_ADM5    41		/*    LSI ADM 5 */
 #define     TT_VTNT    42       /*    Microsoft NT Virtual Terminal */
-#define     TT_MAX   TT_VTNT
+#define     TT_REGENT25 43      /*    ADDS Regent 25 (ADDS25) */
+#define     TT_MAX   TT_REGENT25
 #define     TT_VT420   96	/*    DEC VT-420 */
 #define     TT_VT520   97	/*    DEC VT-520/525 */
 #define     TT_XTERM   98   /*    XTerm */
 #define     TT_TEK40   99	/*    Tektronix 401x */
+/* Other ADDS Regent terminals - not currently emulated */
+#define     TT_REGENT20 100     /* ADDS Regent 20  */
+#define     TT_REGENT40 101     /* ADDS Regent 40  */
+#define     TT_REGENT40P 102    /* ADDS Regent 40+ */
+#define     TT_REGENT60 103     /* ADDS Regent 60  */
+#define     TT_REGENT100 104    /* ADDS Regent 100 */
+#define     TT_REGENT200 105    /* ADDS Regent 200 */
 #define     TT_KBM_EMACS   TT_MAX+1
 #define     TT_KBM_HEBREW  TT_MAX+2
 #define     TT_KBM_RUSSIAN TT_MAX+3
@@ -1133,6 +1141,19 @@ struct stringint {			/* String and (wide) integer */
 #define ISVTNT(x)  (x == TT_VTNT)
 #define ISADM3A(x) (x == TT_ADM3A)
 #define ISADM5(x)  (x == TT_ADM5)
+/* ADDS Regent Terminals */
+#define ISREGENT20(x) (x == TT_REGENT20)
+#define ISREGENT25(x) (x == TT_REGENT25 || x == TT_REGENT20)
+#define ISREGENT40(x) (x == TT_REGENT40 || x == TT_REGENT25 \
+                            || x == TT_REGENT20)
+#define ISREGENT40P(x) (x == TT_REGENT40P || x == TT_REGENT40 \
+                            || x == TT_REGENT25 || x == TT_REGENT20)
+#define ISREGENT60(x) (x == TT_REGENT60 || x == TT_REGENT40P \
+                            || x == TT_REGENT40 || x == TT_REGENT25 \
+                            || x == TT_REGENT20)
+#define ISREGENT200(x) (x == TT_REGENT60 || x == TT_REGENT40P \
+                            || x == TT_REGENT40 || x == TT_REGENT25 \
+                            || x == TT_REGENT20 || x == TT_REGENT200)
 #define ISXTERM(x) (x == TT_XTERM)
 #define ISK95(x)   (x == TT_K95)
 #endif /* OS2 */